+++ backend/package.json
@@ -19,7 +19,7 @@
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
     "test:e2e": "jest --config ./test/jest-e2e.json",
     "db:reset": "ts-node -r tsconfig-paths/register ./src/cli.ts db:reset",
-    "scraper:inject:structured:test": "tsx ./src/util/scraper/inject-structured-test.ts"
+    "scraper:inject:structured:test": "ts-node -r tsconfig-paths/register ./src/cli.ts scraper:inject:structured:test"
   },
   "dependencies": {
     "@nestjs/common": "^11.0.1",
