+++ backend/src/util/scraper/scraper.module.ts
@@ -7,6 +7,7 @@ import { LoggerModule } from "../logger/logger.module";
 import { ScraperDbModule } from "./database/scraper.db.module";
 import { MediaStorageModule } from "../../media/media_storage/media_storage.module";
 import { CsvInjectionService } from "./csv/csv-injection.service";
+import { InjectCsvEngine } from "./csv/inject-csv.engine";
 
 @Module({
   providers: [
@@ -14,6 +15,7 @@ import { CsvInjectionService } from "./csv/csv-injection.service";
     ScraperEngine,
     ScraperRunner,
     CsvInjectionService,
+    InjectCsvEngine,
   ],
   imports: [LanguageModule, LoggerModule, ScraperDbModule, MediaStorageModule],
   exports: [ScraperEngine, ScraperRunner, ScraperService, CsvInjectionService],
